requirements.txt
# Core
ultralytics>=8.0.0
onnxruntime-gpu
onnx
opencv-python
torch
torchvision
pyyaml
pillow
numpy

# Dashboard
flask
flask-socketio
eventlet
